requirements.txt
streamlit
pypdf
sentence-transformers
langchain
qdrant-client
redis
google-cloud-firestore
tqdm
python-dotenv


#GoogleAI 
google.generativeai          

#VertexAI
google-cloud-aiplatform 
vertexai 